+++ examples/forced-read.js
@@ -6,7 +6,7 @@ const delay = millis => new Promise(resolve => setTimeout(resolve, millis));
 
 const forcedRead = async sensor => {
   await sensor.triggerForcedMeasurement();
-  await delay(sensor.typicalMeasurementTime());
+  await delay(sensor.maximumMeasurementTime());
   console.log(await sensor.read());
 };