+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * rename triggerForcedRead to triggerForcedMeasurement
+  * add maximumMeasurementTime
 
 1.1.0 - Oct 10 2019
 ===================